+++ Changelog.md
@@ -10,6 +10,7 @@
 ### Refactor
 * Simplify logic of showing current log in Server window
 * Use partial for `loading_overlay`
+* Make `unbookAllServers` async
 
 ### Fixes
 * Fix broken `fetch-ip` button in server details